+++ core/admin/PostBack.php
@@ -90,7 +90,9 @@ class PostBack extends BasePostBack {
 			$generator = demopress()->get_generator( $gen_input );
 
 			if ( ! is_wp_error( $generator ) ) {
-				$process = Process::instance( $this->a()->n(), $this->a()->plugin_prefix )->prepare( $generator->settings_for_processing() )->process();
+				$request = isset( $_REQUEST[ $this->a()->n() ] ) ? Sanitize::deep( $_REQUEST[ $this->a()->n() ], 'html' ) : array();
+				$process = Process::instance( $this->a()->n(), $this->a()->plugin_prefix )->prepare( $generator->settings_for_processing() )->process( $request );
+
 				$request = $generator->process_request( $process[ $gen_input ] );
 
 				demopress_gen()->new_task( $gen_input, $request );