+++ core/data/video/Base.php
@@ -10,5 +10,4 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 abstract class Base extends Video {
 	public $scope = 'remote';
-
 }