+++ core/data/video/PixabayCom.php
@@ -12,10 +12,7 @@ class PixabayCom extends Base {
 	/**
 	 * @param array                                      $settings
 	 * @param \Dev4Press\Plugin\DemoPress\Base\Generator $generator
-	 *
-	 * @return mixed
 	 */
 	public function run( $settings = array(), $generator = null ) {
-
 	}
 }