+++ forms/content-dashboard.php
@@ -17,14 +17,14 @@ if ( ! defined( 'ABSPATH' ) ) {
 			?>
 
             <div class="d4p-features-wrapper">
-			<?php
+				<?php
 
-			panel()->include_generic( 'element', 'subpanels', 'blocks', array( 'class' => '_is-settings' ) );
+				panel()->include_generic( 'element', 'subpanels', 'blocks', array( 'class' => '_is-settings' ) );
 
-			?>
+				?>
             </div>
 
-            <?php
+			<?php
 		} else {
 			$generator = demopress()->get_generator( demopress_admin()->subpanel );
 
@@ -79,12 +79,12 @@ if ( ! defined( 'ABSPATH' ) ) {
 					<?php _e( "Generator", "demopress" ); ?>:
                     <strong><?php echo demopress()->get_generator_label( demopress_gen()->type ); ?></strong>
                     <br/><?php _e( "Started", "demopress" ); ?>:
-                    <strong><?php echo date( "c", demopress_gen()->started ); ?></strong>
+                    <strong><?php echo date( 'c', absint( demopress_gen()->started ) ); ?></strong>
 					<?php if ( ! demopress_gen()->is_running() ) { ?><br/><?php _e( "Ended", "demopress" ); ?>:
                         <strong><?php echo date( "c", demopress_gen()->ended ); ?></strong><?php } ?>
                 </div>
                 <div class="demopress-gen-status">
-                    <pre><?php echo join( D4P_EOL, demopress_gen()->format_log_list() ); ?></pre>
+                    <pre><?php echo join( PHP_EOL, demopress_gen()->format_log_list() ); ?></pre>
                 </div>
                 <p>
 					<?php _e( "This page automatically checks for the generator status every 5 seconds. If the process gets broken on the server side, the plugin will attempt auto-recovery.", "demopress" ); ?>