+++ forms/generators/bbpress.php
@@ -1,3 +1,10 @@
+<?php
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
+?>
 <div class="d4p-group d4p-group-information">
     <h3><?php _e( "bbPress Generator", "demopress" ); ?></h3>
     <div class="d4p-group-inner">