+++ history.md
@@ -2,6 +2,21 @@
 
 ## Changelog
 
+### Version: 1.6.1 / may 15 2022
+
+* **new** tested with the WordPress 6.0
+
+### Version: 1.6 / april 13 2022
+
+* **new** posts generators: support for the post status
+* **new** bbpress generators: support for the post status
+
+### Version: 1.5 / october 7 2021
+
+* **new** generator: generate WooCommerce products
+* **edit** for content generator show notice for taxonomies with no terms
+* **edit** for content generator show only public taxonomies
+
 ### Version: 1.4 / august 3 2021
 
 * **new** images placeholder generator: include rectangles